# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "CarPaint AI API",
        "title": "CarPaint AI",
        "contact": {},
        "license": {
            "name": "GPLv3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        }
    },
    "paths": {
        "/openai/{+}": {
            "post": {
                "tags": [
                    "OpenAI"
                ],
                "summary": "OpenAI 转发服务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务名称",
                        "name": "+",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Content-Type",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/add": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "新增 API Key",
                "responses": {}
            }
        },
        "/user/delete/{apiKey}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "apiKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.UserInfo"
                        }
                    }
                }
            }
        },
        "/user/recharge": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "充值",
                "parameters": [
                    {
                        "description": "RechargeDto",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RechargeDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.UserInfo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.RechargeDto": {
            "type": "object",
            "required": [
                "amount",
                "recharge_id",
                "target_key"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "recharge_id": {
                    "type": "string"
                },
                "target_key": {
                    "type": "string"
                }
            }
        },
        "entity.UserInfo": {
            "type": "object",
            "properties": {
                "models": {
                    "description": "可用模型",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "recharge_id_records": {
                    "description": "充值记录",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total_tokens": {
                    "type": "number"
                },
                "used_tokens": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Admin-Token",
            "in": "header"
        }
    }
}